feat(ai): introduce embedding dimension normalization and utility methods
- Add `EmbeddingVector` utility class for consistent embedding normalization
- Replace hardcoded embedding dimensions with `MAX_EMBEDDING_DIMENSIONS` constant
- Update OpenAI and Google embedding clients to use new normalization methods
- Remove deprecated static embedding utility methods from `EmbeddingModelClient`
- Ensure all embeddings are normalized to 2048 dimensions for consistency
